+++ Editor/AddressableImporter.cs
@@ -9,15 +9,23 @@ using System.Linq;
 using System.IO;
 using System.Text.RegularExpressions;
 using UnityEditor.AddressableAssets.Settings.GroupSchemas;
+using UnityEditor.Experimental.SceneManagement;
 
 public class AddressableImporter : AssetPostprocessor
 {
     static void OnPostprocessAllAssets(string[] importedAssets, string[] deletedAssets, string[] movedAssets, string[] movedFromAssetPaths)
     {
+        // Skip if all imported and deleted assets are addressables configurations.
+        var isConfigurationPass =
+            (importedAssets.Length > 0 && importedAssets.All(x => x.StartsWith("Assets/AddressableAssetsData"))) &&
+            (deletedAssets.Length > 0 && deletedAssets.All(x => x.StartsWith("Assets/AddressableAssetsData")));
+        if (isConfigurationPass) {
+            return;
+        }
         var settings = AddressableAssetSettingsDefaultObject.Settings;
         if (settings == null)
         {
-            Debug.LogWarningFormat("[Addressables] settings file not found.\nPlease go to Menu/Window/Asset Management/Addressables/Groups, then click 'Create Addressables Settings' button.");
+            Debug.LogWarningFormat("[Addressables] settings file not found.\nPlease go to Menu/Window/Asset Management/Addressables, then click 'Create Addressables Settings' button.");
             return;
         }
         var importSettings = AddressableImportSettings.Instance;
@@ -25,41 +33,86 @@ public class AddressableImporter : AssetPostprocessor
             Debug.LogWarningFormat("[AddressableImporter] import settings file not found.\nPlease go to Assets/AddressableAssetsData folder, right click in the project window and choose 'Create > Addressable Assets > Import Settings'.");
             return;
         }
-        else if (importSettings.rules == null || importSettings.rules.Count == 0)
+        if (importSettings.rules == null || importSettings.rules.Count == 0)
             return;
-        var entriesAdded = new List<AddressableAssetEntry>();
-        foreach (string assetPath in importedAssets)
+
+        var dirty = false;
+
+        // Apply import rules.
+        var prefabStage = PrefabStageUtility.GetCurrentPrefabStage();
+        foreach (var importedAsset in importedAssets)
         {
-            foreach (var rule in importSettings.rules)
-            {
-                if (rule.Match(assetPath))
+            if (prefabStage == null || prefabStage.prefabAssetPath != importedAsset) // Ignore current editing prefab asset.
+                dirty |= ApplyImportRule(importedAsset, null, settings, importSettings);
+        }
+
+        for (var i = 0; i < movedAssets.Length; i++)
+        {
+            var movedAsset = movedAssets[i];
+            var movedFromAssetPath = movedFromAssetPaths[i];
+            if (prefabStage == null || prefabStage.prefabAssetPath != movedAsset) // Ignore current editing prefab asset.
+                dirty |= ApplyImportRule(movedAsset, movedFromAssetPath, settings, importSettings);
+        }
+
+        foreach (var deletedAsset in deletedAssets)
+        {
+            if (TryGetMatchedRule(deletedAsset, importSettings, out var matchedRule)) {
+                var guid = AssetDatabase.AssetPathToGUID(deletedAsset);
+                if (!string.IsNullOrEmpty(guid) && settings.RemoveAssetEntry(guid))
                 {
-                    var entry = CreateOrUpdateAddressableAssetEntry(settings, importSettings, rule, assetPath);
-                    if (entry != null)
-                    {
-                        entriesAdded.Add(entry);
-                        if (rule.HasLabel)
-                            Debug.LogFormat("[AddressableImporter] Entry created/updated for {0} with address {1} and labels {2}", assetPath, entry.address, string.Join(", ", entry.labels));
-                        else
-                            Debug.LogFormat("[AddressableImporter] Entry created/updated for {0} with address {1}", assetPath, entry.address);
-                    }
+                    dirty = true;
+                    Debug.LogFormat("[AddressableImporter] Entry removed for {0}", deletedAsset);
                 }
             }
         }
-        if (entriesAdded.Count > 0)
-        {
-            settings.SetDirty(AddressableAssetSettings.ModificationEvent.EntryMoved, entriesAdded, true);
+
+        if (dirty) {
             AssetDatabase.SaveAssets();
-        }
-        if (importSettings.removeEmtpyGroups)
-        {
-            settings.groups.RemoveAll(_ => _.entries.Count == 0 && !_.IsDefaultGroup());
         }
     }
 
     static AddressableAssetGroup CreateAssetGroup<SchemaType>(AddressableAssetSettings settings, string groupName)
     {
         return settings.CreateGroup(groupName, false, false, false, new List<AddressableAssetGroupSchema> { settings.DefaultGroup.Schemas[0] }, typeof(SchemaType));
+    }
+
+    static bool ApplyImportRule(
+        string assetPath,
+        string movedFromAssetPath,
+        AddressableAssetSettings settings,
+        AddressableImportSettings importSettings)
+    {
+        var dirty = false;
+        if (TryGetMatchedRule(assetPath, importSettings, out var matchedRule))
+        {
+            // Apply the matched rule.
+            var entry = CreateOrUpdateAddressableAssetEntry(settings, importSettings, matchedRule, assetPath);
+            if (entry != null)
+            {
+                if (matchedRule.HasLabel)
+                    Debug.LogFormat("[AddressableImporter] Entry created/updated for {0} with address {1} and labels {2}", assetPath, entry.address, string.Join(", ", entry.labels));
+                else
+                    Debug.LogFormat("[AddressableImporter] Entry created/updated for {0} with address {1}", assetPath, entry.address);
+            }
+
+            dirty = true;
+        }
+        else
+        {
+            // If assetPath doesn't match any of the rules, try to remove the entry.
+            // But only if movedFromAssetPath has the matched rule, because the importer should not remove any unmanaged entries.
+            if (!string.IsNullOrEmpty(movedFromAssetPath) && TryGetMatchedRule(movedFromAssetPath, importSettings, out matchedRule))
+            {
+                var guid = AssetDatabase.AssetPathToGUID(assetPath);
+                if (settings.RemoveAssetEntry(guid))
+                {
+                    dirty = true;
+                    Debug.LogFormat("[AddressableImporter] Entry removed for {0}", assetPath);
+                }
+            }
+        }
+
+        return dirty;
     }
 
     static AddressableAssetEntry CreateOrUpdateAddressableAssetEntry(
@@ -118,6 +171,23 @@ public class AddressableImporter : AssetPostprocessor
         return entry;
     }
 
+    static bool TryGetMatchedRule(
+        string assetPath,
+        AddressableImportSettings importSettings,
+        out AddressableImportRule rule)
+    {
+        foreach (var r in importSettings.rules)
+        {
+            if (!r.Match(assetPath))
+                continue;
+            rule = r;
+            return true;
+        }
+
+        rule = null;
+        return false;
+    }
+
     /// <summary>
     /// Find asset group by given name. Return default group if given name is null.
     /// </summary>
@@ -152,10 +222,55 @@ public class AddressableImporter : AssetPostprocessor
     /// </summary>
     public class FolderImporter
     {
-        [MenuItem("Assets/AddressablesImporter: Check Folder(s)")]
-        private static void CheckFolders()
+        /// <summary>
+        /// Reimporter folders.
+        /// </summary>
+        /// <param name="settings">Reference to the <see cref="AddressableAssetSettings"/></param>
+        public static void ReimportFolders(IEnumerable<String> assetPaths)
         {
-            HashSet<string> filesToImport = new HashSet<string>();
+            HashSet<string> pathsToImport = new HashSet<string>();
+            foreach (var assetPath in assetPaths)
+            {
+                if (Directory.Exists(assetPath))
+                {
+                    // Add the folder itself.
+                    pathsToImport.Add(assetPath.Replace('\\', '/'));
+                    // Add sub-folders.
+                    var dirsToAdd = Directory.GetDirectories(assetPath, "*", SearchOption.AllDirectories);
+                    foreach (var dir in dirsToAdd)
+                    {
+                        // Filter out .dirname and dirname~, those are invisible to Unity.
+                        if (!dir.StartsWith(".") && !dir.EndsWith("~"))
+                        {
+                            pathsToImport.Add(dir.Replace('\\', '/'));
+                        }
+                    }
+                    // Add files.
+                    var filesToAdd = Directory.GetFiles(assetPath, "*", SearchOption.AllDirectories);
+                    foreach (var file in filesToAdd)
+                    {
+                        // Filter out meta and DS_Store files.
+                        if (!file.EndsWith(".meta") && !file.EndsWith(".DS_Store"))
+                        {
+                            pathsToImport.Add(file.Replace('\\', '/'));
+                        }
+                    }
+                }
+            }
+            if (pathsToImport.Count > 0)
+            {
+                Debug.Log($"AddressablesImporter: Found {pathsToImport.Count} asset paths...");
+                OnPostprocessAllAssets(pathsToImport.ToArray(), new string[0], new string[0], new string[0]);
+            }
+        }
+
+        /// <summary>
+        /// Allows assets within the selected folder to be checked agains the Addressable Importer rules.
+        /// </summary>
+        [MenuItem("Assets/AddressablesImporter: Check Folder(s)")]
+        private static void CheckFoldersFromSelection()
+        {
+            List<string> assetPaths = new List<string>();
             // Folders comes up as Object.
             foreach (UnityEngine.Object obj in Selection.GetFiltered(typeof(UnityEngine.Object), SelectionMode.Assets))
             {
@@ -163,32 +278,15 @@ public class AddressableImporter : AssetPostprocessor
                 // Other assets may appear as Object, so a Directory Check filters directories from folders.
                 if (Directory.Exists(assetPath))
                 {
-                    var filesToAdd = Directory.GetFiles(assetPath, "*", SearchOption.AllDirectories);
-                    foreach (var file in filesToAdd)
-                    {
-                        // If Directory.GetFiles accepted Regular Expressions, we could filter the metas before iterating.
-                        if (!file.EndsWith(".meta") && !file.EndsWith(".DS_Store"))
-                        {
-                            filesToImport.Add(file.Replace('\\', '/'));
-                        }
-                    }
+                    assetPaths.Add(assetPath);
                 }
             }
-
-            if (filesToImport.Count > 0)
-            {
-                Debug.Log($"AddressablesImporter: Found {filesToImport.Count} assets...");
-                OnPostprocessAllAssets(filesToImport.ToArray(), null, null, null);
-            }
-            else
-            {
-                Debug.Log($"AddressablesImporter: No files to reimport");
-            }
+            ReimportFolders(assetPaths);
         }
 
         // Note that we pass the same path, and also pass "true" to the second argument.
         [MenuItem("Assets/AddressablesImporter: Check Folder(s)", true)]
-        private static bool ValidateCheckFolders()
+        private static bool ValidateCheckFoldersFromSelection()
         {
             foreach (UnityEngine.Object obj in Selection.GetFiltered(typeof(UnityEngine.Object), SelectionMode.Assets))
             {